refactor: updated logic for submit button
