+++ ddp-workspace/projects/ddp-singular/src/app/components/activity/activity.component.html
@@ -175,7 +175,7 @@
       (mouseenter)="mouseEnterOnSubmit()"
     >
       <mat-icon *ngIf="displayAgreeIcon">done</mat-icon>
-      {{ (isLastStep && isLastActivity ? submitText : 'SDK.NextButton') | translate }}
+      {{ (isLastStep && isLastOfActivities ? submitText : 'SDK.NextButton') | translate }}
     </button>
 
     <button *ngIf="isReadonly()" class="button button--primary pull-right activity-action" (click)="close()">
