+++ ddp-workspace/projects/ddp-singular/src/app/components/pages/survey/survey.component.html
@@ -6,7 +6,7 @@
         [studyGuid]="studyGuid"
         (submit)="onSubmit($event)"
         [activityGuid]="instanceGuid"
-        [isLastActivity]="isLastActivity"
+        [isLastOfMultipleActivities]="isLastOfMultipleActivities"
         *ngIf="!isFetchingActivities"
       ></app-activity>
     </div>
