+++ ddp-workspace/projects/ddp-singular/src/app/utils/index.ts
@@ -1,1 +1,2 @@
 export * from './helpers';
+export * from './types';
