+++ ddp-workspace/projects/ddp-singular/src/app/utils/types.ts
@@ -1,0 +1,13 @@
+export interface RenderActivity {
+  i18nKey: string | null;
+  isCurrent?: boolean;
+  isComplete?: boolean;
+}
+
+export enum RenderActivityKey {
+  Consent = 'CONSENT',
+  About = 'ABOUT',
+  MedicalReleaseForm = 'MEDICAL_RELEASE_FORM',
+  MedicalRecordUpload = 'MEDICAL_RECORD_UPLOAD',
+  PatientSurvey = 'PATIENT_SURVEY',
+}
